rm moves things to StriaTrash
- Stupid behavior, but will fix it later. Probably make a -t switch
  that lets you move the content to a folder instead of deleting.
  Whatever.

+++ cmis/rm.py
@@ -11,14 +11,23 @@ rm_cmd = util.create_command()
 rm_cmd.description = 'Delete a folder or document object from a repository.'
 
 def do_rm(context, obj):
-    try:
+    client = util.create_client(context)
+    repo = client.defaultRepository
+    trash = repo.getObjectByPath('/StriaTrash')
+    print(obj.id)
+
+    props = obj.getProperties()
+
+    if props['cmis:baseTypeId'] == 'cmis:folder':
         for c in obj.getChildren():
             do_rm(context, c)
-    except:
-        pass
-
-    obj.delete()
-
+        obj.delete()
+    elif props['cmis:baseTypeId'] == 'cmis:document':
+        #obj.delete()
+        parents = obj.getObjectParents().getResults()
+        #obj.move(parents[0], trash)
+        doc = repo.getObject(obj.id)
+        doc.move(repo.getObject(parents[0].id), trash)
 
 @rm_cmd.main('path')
 def rm(context, path):
